Move GeoLevel out of shared entities
This is to work around a compilation errors I don't understand and only
see when compiling within the container:

client_1    | ./src/client/components/MapHeader.tsx
client_1    | Module not found: Can't resolve '../../shared/entities' in
'/home/node/app/src/client/components'
